requirements.txt
# --- Core AI & LLM Client ---
openai>=1.10.0
ollama>=0.1.6

# --- Vector Database ---
chromadb>=0.4.22

# --- Search Tools ---
ddgs>=1.0.0

# --- Reranking & Embeddings ---
sentence-transformers>=2.3.1
torch>=2.2.0

# --- Evaluation Framework ---
ragas>=0.1.0
datasets>=2.17.0
langchain>=0.1.0
langchain-community>=0.0.10

# --- UI/UX & Observability ---
chainlit>=1.0.0
langfuse>=2.0.0

# --- Utilities ---
python-dotenv>=1.0.1
pandas>=2.2.0
colorama>=0.4.6
tiktoken>=0.6.0
tqdm>=4.66.0
